fix: add MockPersistentMockRegistry to tests/Fixtures
